customer-sentiment-analyzer/reference/truenas-sentiment-analysis-main/requirements.txt
# Core dependencies
anthropic>=0.39.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.1

# Visualization
matplotlib>=3.7.0
reportlab>=4.0.0
plotly>=5.18.0

# Dashboard
streamlit>=1.29.0

# CLI & UX
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0

# Optional: Slack integration
requests>=2.28.0

# Development
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
